Fix some problems in setup.sh

+++ .devcontainer/setup.sh
@@ -1,6 +1,16 @@
 #!/bin/bash
+python -m venv env
+source env/bin/activate
 poetry install
-python tools/install_bootstrap.py
-cp .env.example .env
+if python tools/install_bootstrap.py --is-installed; then
+    echo "Bootstrap is already installed."
+else
+    python tools/install_bootstrap.py
+fi
+
+if [ ! -f .env ]; then
+    cp .env.example .env
+fi
+
 python manage.py migrate
 python manage.py compilemessages
